Refactor postingsSetsIterator offset fields
